We actually need get_ip(). Script is done now.
get_ip() is actually a bit of a philosophical debate, we could either
hit Linode's API with a DNS resource update every time, or we could
hit the IP check URL every time. Seeing as how the IP check URL
probably takes less resources to run, I'm opting to hit that.

If I wasn't doing that, I could just supply '[remote_addr]' as the
TARGET in the Linode resource update/create call. Cleaner code,
but it's not as nice. :)